refactor(go-imap): split sync.go into 5 focused files
sync.go was 1684 lines with 25 methods mixing coordinator logic,
per-mailbox sync, flag sync, mailbox discovery, and store insertion.
Split into five files along responsibility boundaries:

- sync.go (307): core types (SyncMode, SyncOptions, SyncResult,
  Syncer), constructors, top-level Sync() coordinator, SyncAccounts()
- sync_discovery.go (232): mailbox discovery via SPECIAL-USE,
  Gmail helpers (isGmail, resolveGmailMailboxes, gmailLabelsToTags),
  pattern matching
- sync_mailbox.go (433): per-mailbox sync loop (syncMailbox),
  connection error detection, MessageID mapping cache, header
  backfill
- sync_flags.go (564): flag sync (syncFlags, syncGmailLabels),
  folder moves, upload/download flag changes, tag mapping helpers
- sync_store.go (193): store insertion (storeInsertMessage,
  extractMessageIDFromBody) plus accountName helper

No behavioral changes — all 25 methods are still on Syncer and
behave identically. Go allows splitting struct methods across
files in the same package, so no access modifier changes were
needed. BUILD.bazel srcs list updated to include the four new
files.

All 15 CLI test targets still pass (including imap_test).
