+++ cli/internal/imap/BUILD.bazel
@@ -9,6 +9,10 @@ go_library(
         "rules.go",
         "state.go",
         "sync.go",
+        "sync_discovery.go",
+        "sync_flags.go",
+        "sync_mailbox.go",
+        "sync_store.go",
     ],
     importpath = "github.com/durian-dev/durian/cli/internal/imap",
     visibility = ["//cli:__subpackages__"],
